Updated openstack/openstack Project: openstack/tempest  299699f66ed61615f03eb8780483ed8282c95ac5
Add comments for test module of tempest.test.idempotent_id

Although the depreciate warning is generated, we can not use
tempest.lib.decorators instead here, because the module is for
testing tempest.test.idempotent_id itself. We can not remove
this test module because some projects still use the depreciate
one. Adding the comments is for people that this depreciate
warning will be still here before all the projects switch to use
tempest.lib.decorators, then we can remove this test module, but
this may need a period of time.


Change-Id: I188f1a6502f42e9eb4794aeea76c3ed06dcd4b3b