Updated openstack/openstack Project: openstack/nova  daf4fa7456bd43d84b97bed44891b7ef2cc8279a
Add functional force live migrate test

Forced live migration now seems to be failing, as after the movement of
the VM, the allocations are only present on the source host, and not on
the destination host.
The related assertions are commented out and replaced with wrong ones to
make the test pass.

Related-Bug: #1712008
Change-Id: I6856f57426db6f2f49daea86679b50d5d019fe19