Updated openstack/openstack Project: openstack-infra/puppet-zuul  3ebcb481bb01c1779dd1ee32e4eeef8825b27ab1
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: Ie43d7fd4860f7c0f9593bdd7b5b4cd5b54bcadf4