Updated openstack/openstack Project: openstack/neutron-fwaas  102860c4a4c30bc9e0f6defe98152a613091be33
FW rule applied incorrectly if port specified is a range

When creating a firewall rule with port specified as a range of values,
e.g. [1], conntrack command for deleting current conntrack entries is
applied to the first number in the range, e.g. port #8778 in [1],
instead of the range of ports 8778:9000.

This incorrect behavior occurs because conntrack-tools
does not understand the port as a range of values.
This patch set fixes that issue by following the same method as done
in the netlink implementation in [2].

[1] "neutron firewall-rule-create --protocol tcp --action allow
--ip-version 4 --destination-port 8778:9000 --enabled True"
[2] https://review.openstack.org/#/c/438445/

Closes-Bug: #1702242
Co-Authored-By: Vu Cong Tuan <tuanvc@vn.fujitsu.com>
Change-Id: Ib17db09069a07f35109357d20b67b1acfa85c1a4