Updated openstack/openstack Project: openstack/nova  36fc828c612a0aa5c7bd1363107b84facb98044c
Clarifying node_uuid usage in ironic driver.

Original bug text:
The _get_node() method in the Ironic virt driver has the following
signature and docstring:

    def _get_node(self, node_uuid):
        """Get a node by its UUID."""

However, there are 4 places in the code where this method is called
that passes the string for the nodename instead:

node_is_available()
_node_from_cache()
plug_vifs()
unplug_vifs()

I don't know how any of these are working, but I guess the periodic
update that correctly calls _get_node() with a UUID keeps the node
cache current.
Original bug text END!

This change addresses bug 1704231. In the Ironic driver nodename is
often times used because the inherited class uses nodename for method
parameters. However the ironic driver doesn't actually use the
nodename but uses node_uuid. Driver.py is update by code in some
places and documentation in others to try to make it more clear
that node_uuid is being used without changing the signature of the
methods that are inhereted.

Change-Id: I3b69c7d90ed40816e740a2a609c115312fb466ac
Closes-Bug: #1704231