Updated openstack/openstack Project: openstack/tripleo-heat-templates  c9c4d7e8cddd0f517de22167dbf326ad3dcdea3a
Do not install additional CentOS packages via ceph-ansible

EPEL is required by the Ceph RPM repo maintained at ceph.com but
not by the CentOS SIG repo. The other packages should be already
installed in the infra image.

This change overrides the list of required packages in ceph-ansible
so that it does not install epel-release nor additional unneeded
packages [1].

1. https://github.com/ceph/ceph-ansible/blob/master/group_vars/all.yml.sample#L67

Change-Id: I7b6d2794ebfc0df5b3d5ad9de1d1d55740a38ae0
Closes-Bug: #1712599