Updated openstack/openstack Project: openstack/python-glanceclient  b90236eef83e3d6234cbcdf2ef4dc56bfb5455c7
image-create-via-import fails with ValueError

CLI image-create-via-import fails with ValueError. The reason
is create command returns RequestIdWrapper object and not image
and response.

Further it fails with AttributeError: 'Namespace' object has
no attribute 'import_method'. The reason is do_import_image() call
requires 'import_method' as input which is not provided at the
moment.

Change-Id: Ic4c4d1f3c5d290b584840e8f9047fb53611a5748
Closes-bug: #1711511