Updated openstack/openstack Project: openstack-infra/puppet-openstackid  4c99e995a8ec0cc64ca59c2958ff34981543ae1c
Allow multiple places to install curl

Curl is an important package that we end up needing in a few places. To
avoid puppet unique name conflicts with the package "curl" only define
the curl package resource if it isn't already defined.

This allows us to define it in multiple places that may or may not
overlap depending on how hosts are constructed.

Change-Id: I17c119e81a035916550de44356b9f4badd505c81