Updated openstack/openstack Project: openstack-infra/puppet-etherpad_lite  5aae261ebf5aea1103b3de55d333f2b02cf18b8a
Allow multiple places to install curl

Curl is an important package that we end up needing in a few places. To
avoid puppet unique name conflicts with the package "curl" only define
the curl package resource if it isn't already defined.

This allows us to define it in multiple places that may or may not
overlap depending on how hosts are constructed.

Change-Id: I8b346cbfb908c5d2094c5e46ff008ef735b0e903