Updated openstack/openstack Project: openstack-infra/puppet-pip  1c6e7f380e5d81f31863279e96ceeb850c8a9e9e
Trim the test package list to just shade

We were installing ~24 packages before. Problem with this is it takes
quite a bit of time and is subject to many external failures. Installing
just shade should still be a valid test of significant coverage as it is
a non trivial python package and we use it in quite a few places.

Change-Id: I9cefc83ff280ca00cb434dbe6a1a2357163bc8a6