Updated openstack/openstack Project: openstack/nova  a99e936ff14c3993df1b3168cf2cdb28dcbe49b2
Cleanup allocations in failed prep_resize

During a resize/migration, the scheduler 'doubles' the
allocations on both the source and destination hosts, which
could be the same host if resizing to the same host.

If prep_resize fails, the destination node allocations were
not getting cleaned up before rescheduling to another host.
If it's a resize to the same host, the doubled allocation
from the scheduler wasn't being subtracted for the single host.

This change cleans up the allocations from the current node
when prep_resize fails. If it's not a resize to the same host,
we're on the destination node already. If it is a resize to
the same host, remove_provider_from_instance_allocation in the
SchedulerReportClient accounts for subtracting the new flavor
from the doubled allocation.

Change-Id: I8e81704518cef8847dc65b70a75cbd5e67f1cd39
Closes-Bug: #1712850