Updated openstack/openstack Project: openstack/nova  42e83eed3e60a9909e9f99575a6ff3a7a8baf3f8
[placement] Disambiguate resource provider conflict message

When creating a resource provider, if the uuid or name (or both)
of the resource provider is already being used a 409 response is
returned. Prior to this change the error message always mentioned
the name of the resource provider, even if it was not the cause of
the duplication. That's misleading and unfriendly.

The new code will list the field and the value of the field that
was duplicated. From looking at the olso_db code that creates a
DBDuplicateEntry it appears that sometimes, but not always, the
exception could list both fields, so allowance for that is made
in the message output and the related gabbi tests.

Because this changes the error message, but not the error response
code, this does not violate the api interoperability guidelines[1],
so I'm willing to say this doesn't require a microversion, and thus
does not require a spec.

[1] http://specs.openstack.org/openstack/api-wg/guidelines/api_interoperability.html
If we do think this is a violation, then the guideline should be
made more explicit.

Change-Id: Ibafbfd8302977a2b4e9250fbb5ada283b69d3c25
Closes-Bug: #1693349