Updated openstack/openstack Project: openstack-dev/devstack  9d7e74e57c73d25e214e0269dbd9d342e8cd960c
Setup logging for nova_cell1.conf

When run in the default superconductor mode, the screen-n-cond-cell1
logs are not formatting in oslo format or colorized like the other
logs. This is because screen-n-super-cond is running using nova.conf
which is configured for oslo format logging with color.

The oslo format logging is also needed to correctly index the logs
from screen-n-cond-cell1 in logstash.

This change simply configures nova_cell*.conf files for logging
like nova.conf.

Change-Id: I44fc11f09bb7283be0b068f5e02a424f3e5dafe2
Closes-Bug: #1713070