Updated openstack/openstack Project: openstack-infra/tripleo-ci  d36f5fedbc2ca30c0bb1c84242a99e62877bf120
Do not try to inject repo into IPA image if we dont fetch it

When adding support to inject changes into the IPA image, we missed
a check for whether we are building images rather than fetching them.

This causes the jobs which build images to fail because they try to
modify an image we did not fetch.

Change-Id: Ife155e904a3118fb5b44419845e8fe321ea11e62
Closes-Bug: 1713079