Updated openstack/openstack Project: openstack/keystone  6b6e1a7322666fb4da871281bb8c58db199c7f49
Call methods with kwargs instead of positionals

If a method is defined with kwargs it should be called with
kwargs. Calling a method that has specified kwargs with positional
arguments is fragile and should be avoided. It leaves the caller
open to breakages if the method signature of the callee ever changes.
If the callee's signature changes, it requires the callers to be
updated, even if they aren't using the new functionality or if the
argument doesn't apply to them.

This change applies that practice to parts of the assignment API.
This will make it easier to reuse the existing assignment driver to
implement global role assignments.

bp global-roles

Change-Id: I0a228ec25cd88f456cacd5824082f1b322e0262d