Updated openstack/openstack Project: openstack-infra/elastic-recheck  a8eaf1ab6950b9d02f89e7101938d0d9473fdb1e
Track issues with invalid repo files, since baseurl is empty

So that jobs failed because of an invalid repo file, since baseurl is
empty, are classified as a such and not noise in the uncategorized
list, lets track these generically.
We can also have specific bugs tracked in addition to this one (like
delorean repos for tripleo).

Change-Id: I2830c2bfba5000702c7b7f0e6d180143c5ebd40c
Related-bug: #1710605
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>