Updated openstack/openstack Project: openstack/python-manilaclient  49c21196bcdb58469d3db3567466814e9f89ea8f
doc-migration: new directory layout

This patch introduces a new directory layout
in doc/source in conformance with the OpenStack
manuals project migration spec [1], moves the
existing content in manila/doc/source into the
new directories, and adjusts index files accordingly.

This is the first step in the migration process
as outlined in the spec.

[1] https://specs.openstack.org/openstack/docs-specs/specs/pike/os-manuals-migration.html

Closes-Bug: #1706181
Change-Id: I8964d066bb838fabbe94239ac108bff884c6ff76
Depends-On: I72be6303f9be77c0461eaacbc61bad8372546fb5
Depends-On: Ia750cb049c0f53a234ea70ce1f2bbbb7a2aa9454