Updated openstack/openstack Project: openstack/os-net-config  314ce2dc8686defe9066a20a4488c8f5811aa352
Add NIC Mapping Reporting Feature

The -i or --interfaces flag will print out a dict of the NIC
mappings if no interfaces are specified, or it will print a
list of the real interface names that match the interfaces
given. If a real interface name is given, it will be returned
in the list without translation only if it is live.

It has always been technically possible to map an interface
using an alias that matches an inactive real interface. This
change preserves that functionality, but will prevent mapping
using the name of an active interface. A warning will be
printed if an alias matches an inactive inteface.

Also change _is_active_nic and _mapped_nics to public functions
since they are called outside of utils.

Change-Id: I74af5391165e10f04800ea05e4204a7d1f74f526
Closes-bug: 1713137