Updated openstack/openstack Project: openstack/openstack-manuals  990a7cf2cb3feaf4d651b13557f2d08a9391cecc
use whereto to test redirect rules

Add a file containing the tests for all of the redirect rules and set
up whereto to run during the check. The test file is a template, just
like the htaccess file, to ensure we auto-generate the rules for all
of the various project redirects.

Fix a few issues with the template generator that ended up producing
duplicate or useless rules.

Change-Id: I0d306149dc33f64d44d9482fab02101bd1f6d7f5
Signed-off-by: Doug Hellmann <doug@doughellmann.com>