Updated openstack/openstack Project: openstack/tempest  55c34fdd4b50fce7ecc29d725a9fa18c44c8af8d
Fix for resetting volume status

Reset volume status API is a volume action and its return code is 202,
which means it is not a sync procedure. Although the API is mainly a
DB operation and the volume status will be changed quickly, it is not
stable to verify the volume status immediately after calling the API.
It is more suitable to use 'waiters' function to wait for the volume
status to be changed.

Change-Id: I0421acd562821af268ca5b20e0424455d71da338