Updated openstack/openstack Project: openstack/nova  ea0e41a880e2656aacf1aef805ef4c0a0b754f20
Don't warn on expected network-vif-unplugged events

If we're getting a network-vif-unplugged event because we're
deleting an instance, and part of that is unplugging vifs,
then we don't need to log a warning about it since it's expected.

This could be true of other tasks like migrations, but we're
starting small here. So for now, if we log a warning, this change
also includes the vm_state and task_state in the warning to help
us discover what else we can expect.

Change-Id: I080075fa543bd1ca4fa57580583bda6d3847bf9c
Closes-Bug: #1691615