Updated openstack/openstack Project: openstack/nova  ba3c809cd8db4b1e4ce4b91d49f358107c44ed42
VMware: Do not check if folder already exists in vCenter

When creating a project folder, we iterate the existing folders to check
if it is already created. This may take long time if there are many
project folders. Given that project folders are never deleted, we should
avoid iterating the existing folders. Instead, we should always create
the project folder and handle DuplicateName exceptions.

Change-Id: If734e22af223836b424561e8e9ca65e6c39fe5c2
Closes-Bug: #1627693