Updated openstack/openstack Project: openstack-infra/project-config  2a15fd9b46d894cbcd357d5fc90e5a863f5ed6e4
Add propose-update-constraints job

After we release to PyPI, we want to send out update-constraints.

We don't need to rewrite all the scripts involved with this for now, so
update them to call each other using relative not absolute paths.

Then, copy the scripts dir from the trusted checkout on the executor
to the homedir on the remote host so that we can execute the appropriate
script. Local paths are relative to the location of the playbook.

Change-Id: I080f71fefe79f7faa581fae91eac74384cef4050