Updated openstack/openstack Project: openstack/oslo.messaging  ba30a3067d7db81cc864ae9e52b2761df44efddd
Ensure RPC endpoint target attribute is correct

An endpoint can have an optional 'target' attribute which is used to
filter the callable endpoint by the target.version or .namespace
attributes.  Therefore 'target' is reserved and attempting to use an
endpoint that overrides the target attribute (say with a function
call) should fail with a TypeError.

Change-Id: I0bbf9fca0ecbe71efa87c9613ffd32eb718f2c0e
Closes-Bug: #1709131