Updated openstack/openstack Project: openstack/heat  cd3f84e2d199c31aef4526d29103067a2d9919bf
Show correct version of data in convergence resource list

In convergence, there can be multiple versions of the same resource
present in a stack at one time. The fix for bug 1301320 was supposed to
make data for all versions available from the resource list API. And
while it does go through all of the resources in the database and create
Resource objects for each one, they were each created with the correct
version of the resource definition but DB data from the latest resource.
(Only the name is passed to the Resource object, it has to figure out
which database row to load data from itself - and it will choose the
same one every time.)

With this change, we always load the correct DB data into the
newly-created Resource object.

Change-Id: I6b9d1b86b3dbf767bccebddd78275bbf0933029a
Closes-Bug: #1704194