Updated openstack/openstack Project: openstack/nova  b74115c2f55f5340af60d1ec2321d91c4098fb21
Improve assertJsonEqual error reporting

In the notification sample tests big json structures are asserted.
If the difference is only in the number of keys in a dict or in
the number of items in a list then the current assert only states
the number of items. If the list or dict is big it is quite hard to
see what is the missing key or item.

This patch adds more descriptive output to the error message if the
json structures are not equal.

Error message without this patch: http://paste.openstack.org/show/616525/
Error message with this patch: http://paste.openstack.org/show/616526/

Change-Id: Iec15c9ea863db6bbe18afd64bd00a0780798cdf8