Updated openstack/openstack Project: openstack/heat  fe02bb8eb71eec66d05e55ae07de629497c1a5cb
Use stored properties values in actions

During update, delete and signal operations, we set the properties of a
resource to the values stored in the database when they were last created
or updated, so that we don't resolve intrinsic functions to new values
caused by possible changes in live attribute values. In fact, since we
started requiring the resources' node data to be pre-populated into the
StackDefinition, we will get only placeholder values for attributes and
resource IDs if we tried to resolve live functions.

This change does the same for the suspend, resume, snapshot, and check
actions. Should any plugin's handler function for any of those actions read
the resource's properties (I'm not sure that any do) then they will now get
the stored values.

Change-Id: If6b6f66877e23b8538a97aa339357ca1a2a29276