Updated openstack/openstack Project: openstack/glance  e1321298e7899d0d0d6b96243935454c427e2464
Change variable used by log message

The code was changing the status of a task and then logging an
info message about the change using the already-changed variable.
This patch saves the old status before the task status is modified
so that the status transition can be accurately reported in the log.

Change-Id: I84fef1d710f24f601971a348f53d74b6a6d6a2f9
Closes-bug: #1713740