Updated openstack/openstack Project: openstack-infra/devstack-gate  62557046fb8f1578af25584a99c4bcaacfdce3d1
Add the private_ipv4 hostvar to primary and sub nodes

This will allow us to reliably set up the bridge on the right IPs
in the follow up network overlay patch.
The private_ipv4 key is inside the nodepool dict to be forward
compatible with the hostvars provided by zuul v3.

Co-Authored-By: Clark Boylan <cboylan@sapwetik.org>
Co-Authored-By: David Moreau Simard <dmsimard@redhat.com>

Change-Id: I34663fc482d71781d1e05455edb18c4a86aa5afb